HEEDLS-383 Move CentreConfiguration files into folders
