+++ DigitalLearningSolutions.Data.Tests/Helpers/SectionContentTestHelper.cs
