+++ DigitalLearningSolutions.Web.Tests/TestHelpers/SectionContentHelper.cs
