+++ DigitalLearningSolutions.Web.Tests/ViewModels/TrackingSystem/CentreConfigurationViewModelTests.cs
@@ -3,6 +3,7 @@
     using System;
     using DigitalLearningSolutions.Data.Tests.Helpers;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.CentreConfiguration;
     using FluentAssertions;
     using FluentAssertions.Execution;
     using NUnit.Framework;
