+++ DigitalLearningSolutions.Web/Controllers/TrackingSystem/CentreConfiguration/RegistrationPromptsController.cs
@@ -1,46 +1,27 @@
-namespace DigitalLearningSolutions.Web.Controllers.TrackingSystem
+﻿namespace DigitalLearningSolutions.Web.Controllers.TrackingSystem.CentreConfiguration
 {
-    using System;
     using System.Collections.Generic;
     using System.Linq;
-    using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Services;
     using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.CentreConfiguration;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.AspNetCore.Mvc.ModelBinding;
 
     [Authorize(Policy = CustomPolicies.UserCentreAdminOnly)]
-    [Route("/TrackingSystem/CentreConfiguration")]
-    public class CentreConfigurationController : Controller
+    [Route("/TrackingSystem/CentreConfiguration/RegistrationPrompts")]
+    public class RegistrationPromptsController : Controller
     {
-        private readonly ICentresDataService centresDataService;
         private readonly ICustomPromptsService customPromptsService;
 
-        public CentreConfigurationController
-        (
-            ICentresDataService centresDataService,
-            ICustomPromptsService customPromptsService
-        )
+        public RegistrationPromptsController(ICustomPromptsService customPromptsService)
         {
-            this.centresDataService = centresDataService;
             this.customPromptsService = customPromptsService;
         }
 
         public IActionResult Index()
-        {
-            var centreId = User.GetCentreId();
-
-            var centreDetails = centresDataService.GetCentreDetailsById(centreId);
-
-            var model = new CentreConfigurationViewModel(centreDetails);
-
-            return View("Index", model);
-        }
-
-        [Route("RegistrationPrompts")]
-        public IActionResult RegistrationPrompts()
         {
             var centreId = User.GetCentreId();
 
@@ -52,77 +33,7 @@ namespace DigitalLearningSolutions.Web.Controllers.TrackingSystem
         }
 
         [HttpGet]
-        [Route("EditCentreManagerDetails")]
-        public IActionResult EditCentreManagerDetails()
-        {
-            var centreId = User.GetCentreId();
-
-            var centreDetails = centresDataService.GetCentreDetailsById(centreId);
-
-            var model = new EditCentreManagerDetailsViewModel(centreDetails);
-
-            return View(model);
-        }
-
-        [HttpPost]
-        [Route("EditCentreManagerDetails")]
-        public IActionResult EditCentreManagerDetails(EditCentreManagerDetailsViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-
-            var centreId = User.GetCentreId();
-
-            centresDataService
-                .UpdateCentreManagerDetails(centreId, model.FirstName, model.LastName, model.Email, model.Telephone);
-
-            return RedirectToAction("Index");
-        }
-
-        [HttpGet]
-        [Route("EditCentreWebsiteDetails")]
-        public IActionResult EditCentreWebsiteDetails()
-        {
-            var centreId = User.GetCentreId();
-
-            var centreDetails = centresDataService.GetCentreDetailsById(centreId);
-
-            var model = new EditCentreWebsiteDetailsViewModel(centreDetails);
-
-            return View(model);
-        }
-
-        [HttpPost]
-        [Route("EditCentreWebsiteDetails")]
-        public IActionResult EditCentreWebsiteDetails(EditCentreWebsiteDetailsViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-
-            var centreId = User.GetCentreId();
-
-            centresDataService.UpdateCentreWebsiteDetails
-            (
-                centreId,
-                model.CentrePostcode,
-                model.CentreTelephone,
-                model.CentreEmail,
-                model.OpeningHours,
-                model.CentreWebAddress,
-                model.OrganisationsCovered,
-                model.TrainingVenues,
-                model.OtherInformation
-            );
-
-            return RedirectToAction("Index");
-        }
-
-        [HttpGet]
-        [Route("RegistrationPrompts/{promptNumber}/Edit")]
+        [Route("{promptNumber}/Edit")]
         public IActionResult EditRegistrationPrompt(int promptNumber)
         {
             var centreId = User.GetCentreId();
@@ -134,7 +45,7 @@ namespace DigitalLearningSolutions.Web.Controllers.TrackingSystem
         }
 
         [HttpPost]
-        [Route("RegistrationPrompts/{promptNumber}/Edit")]
+        [Route("{promptNumber}/Edit")]
         public IActionResult EditRegistrationPrompt(EditRegistrationPromptViewModel model, string action)
         {
             if (action.StartsWith("delete") && TryGetAnswerIndexFromAction(action, out var index))
@@ -221,7 +132,7 @@ namespace DigitalLearningSolutions.Web.Controllers.TrackingSystem
 
             customPromptsService.UpdateCustomPromptForCentre(User.GetCentreId(), model.PromptNumber, model.Mandatory, model.OptionsString);
 
-            return RedirectToAction("RegistrationPrompts");
+            return RedirectToAction("Index");
         }
 
         private void IgnoreAddNewAnswerValidation()
