+++ DigitalLearningSolutions.Web/Views/TrackingSystem/CentreConfiguration/EditCentreWebsiteDetails.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IConfiguration Configuration
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.CentreConfiguration
 @using Microsoft.Extensions.Configuration
 @model EditCentreWebsiteDetailsViewModel
