+++ DigitalLearningSolutions.Web/Views/TrackingSystem/CentreConfiguration/RegistrationPrompts/EditRegistrationPrompt.cshtml
@@ -1,4 +1,5 @@
 ﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.CentreConfiguration
 @model EditRegistrationPromptViewModel
 
 @{
